requirements/base.txt
celery==5.1.2
django==3.1.13
django-admin-sortable2==1.0
django-anymail==8.4
django-celery-results==2.2.0
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-extensions==3.0.3
django-honeypot==1.0.1
django-user-agents==0.4.0
djangorestframework==3.12.4
drf-jwt==1.19.0
factory_boy==3.2.0
flower==1.0.0
git+git://github.com/YPCrumble/django-reflinks.git@trim_next_parameter#egg=django-reflinks
html2text==2020.1.16
psycopg2==2.8.6
pylibmc==1.6.1
sentry-sdk==1.3.1
stripe==2.63.0
